+++ generators/rest-adapter-core/src/main/java/uk/gov/justice/services/adapter/rest/application/CommonProviders.java
@@ -3,6 +3,7 @@ package uk.gov.justice.services.adapter.rest.application;
 import uk.gov.justice.services.adapter.rest.cors.CorsFeature;
 import uk.gov.justice.services.adapter.rest.interceptor.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.adapter.rest.mapper.BadRequestExceptionMapper;
+import uk.gov.justice.services.adapter.rest.mapper.ForbiddenRequestExceptionMapper;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -19,6 +20,7 @@ public class CommonProviders {
     public Set<Class<?>> providers() {
         final HashSet<Class<?>> classes = new HashSet<>();
         classes.add(BadRequestExceptionMapper.class);
+        classes.add(ForbiddenRequestExceptionMapper.class);
         classes.add(JsonSchemaValidationInterceptor.class);
         classes.add(CorsFeature.class);
         return classes;