+++ generators/rest-adapter-core/src/test/java/uk/gov/justice/services/adapter/rest/application/CommonProvidersTest.java
@@ -3,9 +3,10 @@ package uk.gov.justice.services.adapter.rest.application;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 
-import uk.gov.justice.services.adapter.rest.interceptor.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.adapter.rest.cors.CorsFeature;
+import uk.gov.justice.services.adapter.rest.interceptor.JsonSchemaValidationInterceptor;
 import uk.gov.justice.services.adapter.rest.mapper.BadRequestExceptionMapper;
+import uk.gov.justice.services.adapter.rest.mapper.ForbiddenRequestExceptionMapper;
 
 import java.util.Set;
 
@@ -16,6 +17,6 @@ public class CommonProvidersTest {
     @Test
     public void shouldReturnAllCommonProviders() throws Exception {
         Set<Class<?>> providers = new CommonProviders().providers();
-        assertThat(providers, containsInAnyOrder(BadRequestExceptionMapper.class, JsonSchemaValidationInterceptor.class, CorsFeature.class));
+        assertThat(providers, containsInAnyOrder(BadRequestExceptionMapper.class, ForbiddenRequestExceptionMapper.class, JsonSchemaValidationInterceptor.class, CorsFeature.class));
     }
 }